requirements.txt
# Core AI Libraries
anthropic>=0.40.0
openai>=1.54.0

# Vector Database
chromadb>=0.5.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# API & Web
requests>=2.31.0
httpx>=0.27.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Security & Validation
validators>=0.28.0

# Concurrency
aiohttp>=3.9.0

# Logging
colorlog>=6.8.0

# Development
ipython>=8.12.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
